Add trait
